Convert Cloudwatch to a custom resource
