+++ libraries/cloudwatch.rb
@@ -4,6 +4,7 @@ module Opscode
   module Aws
     module CloudWatch
       include Opscode::Aws::Ec2
+
       def cwh
         require 'aws-sdk'
         Chef::Log.debug('Initializing the CloudWatch Client')
