+++ test/fixtures/cookbooks/aws_test/recipes/cloudwatch_alarm.rb
@@ -1,16 +1,13 @@
 aws_cloudwatch 'kitchen_test_alarm' do
   aws_access_key node['aws_test']['key_id']
   aws_secret_access_key node['aws_test']['access_key']
-  period node['aws_test']['cloudwatch']['period']
-  evaluation_periods node['aws_test']['cloudwatch']['evaluation_periods']
-  threshold node['aws_test']['cloudwatch']['threshold']
-  comparison_operator node['aws_test']['cloudwatch']['comparison_operator']
-  metric_name node['aws_test']['cloudwatch']['metric_name']
-  namespace node['aws_test']['cloudwatch']['namespace']
-  statistic node['aws_test']['cloudwatch']['statistic']
-  dimensions node['aws_test']['cloudwatch']['dimensions']
-  actions_enabled node['aws_test']['cloudwatch']['actions_enabled']
-  alarm_actions node['aws_test']['cloudwatch']['alarm_actions']
+  period 21600
+  evaluation_periods 2
+  threshold 50.0
+  comparison_operator 'LessThanThreshold'
+  metric_name 'CPUUtilization'
+  namespace 'AWS/EC2'
+  statistic 'Maximum'
   action :create
 end
